Primarily worked on two things.
1. Added code that can start to compute the top n words of each topic.
To do this, the c_word matrix needs to be rebuilt again, then some
fancy work on the DenseMatrix needs to happen after. I used this
opportunity to learn how to use aggregate to do this construction
instead of the old method which is in the main body of code presently.

2. Related to this, I started adding tests. Specifically I added tests
for the methods I added (except for comb_op since its trivially an alias
method with a name for _ + _). Tests pass for the new aggregate function
returning c_word. Will refactor the name and improve the code.

Next steps are to pull the code in the primary algorithm into a function
to compare against the new method in a test after finishing the pipeline
for getting useful statistics on most common words at the end of a run.